Add the turn success mechanism (rules not yet implemented)
If a move isn't possible this loop will run again untill there is no more options left.

If this is the case... the turn will pass to the other player.

The Board.move function should include all the rules on how to move pawns across the board. *This is not yet functional!*