Fix the Dice
Turns out there was a probability issue that my first take on the dice didn't account for. This version mimicks the acient dice described in the assignment.

You throw 4 dice, each of those can either be black or white. If you throw all black you get 6, else you get the amount of white that you threw.

A little confusing. But it works.